Update Mekanism.cfg
why the FUCK did the maker of mekanism make a goddamn bronze, BRONZE sword better than diamond.

WHY DOES ADDING TWO BLACK ROCKS MAKE IT DOUBLE DAMAGE?!?!?

WHAT THE FUCK REEEEEE

This brings everything closer to vanilla, and nerfs the shit out of the armour spawnrate on zombies/skeletons etc because it's a massive pain in the dickhole.